Updated HTML and JS
Added date and time in header.
Added JS to date and time so it retrieves the actual date and time of your region.
However, had to create an array with days and months first since the days and months show up as numbers on the system.

Added actual humidity and windspeed in JS so when a city is searched the HTML is changed to the actual humidity and wind speed of that city.

Retrieved the icon url from inside the api data and set up a default icon  image into HTML.

Added HTML into JS so the weather icon inside the api can be retrieved and used instead.

Added weather description into HTML then added weather description to JS so it retrieves the actual description.

Updated a few things in CSS.